dont override DOTNET_ROOT
since patch 7.4 and Dalamud 14, xivlauncher and xivlauncher.core now set DOTNET_ROOT to the runtime directory already (for Dalamud itself), so the existing workaround in Browsingway is no longer necessary. Since the path is different for xlcore users, this also breaks the plugin for them. Rely on DOTNET_ROOT set by xivlauncher instead.